moving store components for playground into subfolder of ClientSDKPlayground
