+++ Workload/app/playground/ClientSDKPlayground/ActionDialog/SharedStateExample.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { useDispatch, useSelector } from "react-redux";
 import { Field, Input, Button } from '@fluentui/react-components';
 import { PanelRightExpand20Regular } from '@fluentui/react-icons';
-import { RootState } from "../../ClientSDKPlaygroundStore/Store";
-import { setLocalSharedStateMessage } from "../../ClientSDKPlaygroundStore/actionDialogSlice";
+import { RootState } from "../Store/Store";
+import { setLocalSharedStateMessage } from "../Store/actionDialogSlice";
 import { callDialogOpen } from "../../../controller/DialogController";
 import { SharedState } from "../../../App";
 import "../../../styles.scss";
