+++ Workload/app/playground/ClientSDKPlayground/UIComponents.tsx
@@ -19,9 +19,9 @@ import { Save24Regular } from "@fluentui/react-icons";
 import { useTranslation } from "react-i18next";
 import "../../styles.scss";
 import { TabContentProps } from "./ClientSDKPlaygroundModel";
-import { RootState } from "../ClientSDKPlaygroundStore/Store";
+import { RootState } from "./Store/Store";
 import { callLanguageGet, callSettingsOnChange } from "../../controller/SettingsController";
-import { setCheckboxChecked, setSampleInput, setSelectedRadio, setSwitchChecked } from "../ClientSDKPlaygroundStore/uiComponentsSlice";
+import { setCheckboxChecked, setSampleInput, setSelectedRadio, setSwitchChecked } from "./Store/uiComponentsSlice";
 
 export function UIComponentsExample(props: TabContentProps) {
   const { workloadClient } = props;
